Add BaseViewModelFragment that handles EN permissions and common errors
